refactor(striker-ui): group manage fence related components
