+++ striker-ui/components/ManageFence/index.tsx
@@ -1,0 +1,3 @@
+import ManageFencePanel from './ManageFencePanel';
+
+export default ManageFencePanel;
